Reference guide

Index

VMD.Vmd

VMD.compare

VMD.n_component

VMD.n_mode

VMD.plot

VMD.vmd

Public API

VMD.Vmd — Type

Vmd{T,S<:Int}

Variational Mode Decompostion 
Return by vmd function.

** Arguments **

signal::Array{T,1}:signal

K::S :modes

signal_d::Array{T,Val{K}}:decomposed signals

freqs::Array{T,Val{K}}:decomposed spectrums

samples::S:signal sampled frequency

VMD.compare — Method

compare(v::Vmd)

visual compare origin signal and sum of decomposed signal

VMD.n_component — Method

n_component(v::Vmd,k)

return No k component decomposed signal

VMD.n_mode — Method

n_mode(v::Vmd,k)
return the k mode frequency

VMD.plot — Method

plot(vmd::Vmd;k=1)

visual the decomposed signals and spectrums

** Argument **

vmd::Vmd : vmd

k::Int : 0-original signal 1-first component enforcement

VMD.vmd — Method

vmd(signal;alpha=100, tau=0, K=3, DC=false, init=1, tol=1e-6,samples=100)

Variational Mode Decomposition
Return Vmd

** Argument **

signal: the time domain signal (1D) to be decomposed

alpha : the balancing parameter of the data-fidelity constraint

tau : time-step of the dual ascent ( pick 0 for noise-slack )

K : the number of modes to be recovered

DC : true if the first mode is put and kept at DC (0-freq)

init : 0 = all omegas start at 0 1 = all omegas start uniformly distributed 2 = all omegas initialized randomly

tol : tolerance of convergence criterion; typically around 1e-6

sample_frequency : samples frequency(eg:100/s)

** Example **

T = 1000;
t = (1:T)/T;
sample_frequency = 1000;

# center frequencies of components
f_1 = 2;
f_2 = 24;
f_3 = 288;

# modes
v_1 = @. cos(2*pi*f_1*t);
v_2 = @. 1/4*(cos(2*pi*f_2*t));
v_3 = @. 1/16*(cos(2*pi*f_3*t));

# composite signal, including noise
f = v_1 + v_2 + v_3 + 0.1*randn(length(v_1));

# some sample parameters for VMD
alpha = 2000;       # moderate bandwidth constraint
tau = 0;            # noise-tolerance (no strict fidelity enforcement)
K = 3;              # 3 modes
DC = false;             # no DC part imposed
init = 1;           # initialize omegas uniformly
tol = 1e-7;

v = vmd(f ; alpha = alpha,tau = tau,K = K,DC = false,init = 1,tol = tol,sample_frequency = sample_frequency)

# plot original signal and spectrum
plot(v;k = 0)

# plot first decomposed component and spectrum
plot(v;k = 1)